chore(irBuidler): Restructure the `builders` module
Towards #60 and #91
