+++ src/internals/ir/builders/index.ts
@@ -1,0 +1,4 @@
+export * from "./ir";
+export * from "./astStore";
+export * from "./tactConfig";
+export * from "./tactStdlib";
